Moves middleware into chatbot dir
