+++ lib/middleware/chatbot/parse-ask-continue-response.js
