+++ lib/middleware/chatbot/parse-ask-signup-response.js
