+++ lib/middleware/chatbot/reply-brain.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const bot = require('../bot.js');
+const bot = require('../../bot.js');
 
 module.exports = function getBotReply() {
   return (req, res, next) => {
