+++ lib/middleware/chatbot/template-brain.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const helpers = require('../helpers');
+const helpers = require('../../helpers');
 
 module.exports = function replyWithBrainResponse() {
   return (req, res, next) => {
