+++ lib/middleware/chatbot/template-gambit.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const gambit = require('../gambit');
+const gambit = require('../../gambit');
 
 module.exports = function gambitTemplate() {
   return (req, res, next) => {
