+++ lib/middleware/chatbot/template-paused.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const logger = require('heroku-logger');
-const front = require('../front');
+const front = require('../../front');
 
 module.exports = function replyNoReply() {
   return (req, res, next) => {
